Move the logins to its own file
